chore: Straighten circular dependencies
