+++ src/buttons/hangmanLessButton.ts
@@ -1,14 +1,12 @@
 import { ButtonBuilder } from '@discordjs/builders';
 import { ButtonStyle } from 'discord.js';
-import {
-	buildEvilHangmanMessage,
-	parseEvilHangmanMessage,
-} from '../evilHangman/evilHangmanMessage';
+import { parseEvilHangmanMessage } from '../evilHangman/parseEvilHangmanMessage';
 
 const customId = 'hangmanLessButton';
 export const hangmanLessButton: Button = {
 	customId,
 	async execute({ interaction, message }): Promise<void> {
+		const { buildEvilHangmanMessage } = await import('../evilHangman/evilHangmanMessage');
 		const game = parseEvilHangmanMessage(message);
 
 		const displayInfo = game.getDisplayInfo();
