+++ src/evilHangman/evilHangmanMessage.test.ts
@@ -1,7 +1,6 @@
 import type { EmbedBuilder } from '@discordjs/builders';
-import type { Message } from 'discord.js';
 import { EvilHangmanWinState } from './evilHangmanGame';
-import { buildEvilHangmanMessage, parseEvilHangmanMessage } from './evilHangmanMessage';
+import { buildEvilHangmanMessage } from './evilHangmanMessage';
 
 describe('evilHangmanMessage', () => {
 	describe('buildEvilHangmanMessage', () => {
@@ -53,40 +52,6 @@ describe('evilHangmanMessage', () => {
 				1
 			);
 			expect(message.components).toBeArrayOfSize(1);
-		});
-	});
-
-	describe('parseEvilHangmanMessage', () => {
-		const field: { value: string | undefined } = {
-			value: 'Remaining Guesses: 4\nWord: -----\nLetters Guessed: ',
-		};
-		const message: Message = {
-			embeds: [
-				{
-					data: {
-						fields: [{ value: '' }, field],
-					},
-				},
-			],
-		} as unknown as Message;
-		beforeEach(() => {
-			field.value = 'Remaining Guesses: 4\nWord: -----\nLetters Guessed: ';
-		});
-
-		test('message text is converted into a game', () => {
-			const game = parseEvilHangmanMessage(message);
-			const displayInfo = game.getDisplayInfo();
-			expect(displayInfo.guessesRemaining).toEqual(4);
-			expect(displayInfo.word).toEqual('-----');
-			expect(displayInfo.guessesSoFar).toBeEmpty();
-		});
-
-		test('incorrect message format throws an error', () => {
-			field.value = 'Incorrectly formatted test string';
-			expect(() => parseEvilHangmanMessage(message)).toThrow();
-
-			field.value = undefined;
-			expect(() => parseEvilHangmanMessage(message)).toThrow();
 		});
 	});
 });
